+++ larq_zoo/sota/__init__.py
@@ -1,4 +1,3 @@
-from larq_zoo.sota.quicknet import QuickNet, QuickNetLarge
-from larq_zoo.sota.quicknet_xl import QuickNetXL
+from larq_zoo.sota.quicknet import QuickNet, QuickNetLarge, QuickNetXL
 
 __all__ = ["QuickNet", "QuickNetLarge", "QuickNetXL"]
